requirements.txt
# Plinian Outreach Webhook - Dependencies
# Railway will auto-detect this file

# Web framework
flask>=3.0.0
gunicorn>=21.0.0

# Notion integration
notion-client>=2.0.0

# Anthropic Claude API
anthropic>=0.39.0

# Google Gmail API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0

# Environment management
python-dotenv>=1.0.0
